rename: index.tsx로 파일명 수정
